# cmp_backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Consent Management Platform",
    "description": "\nConsent Management Platform.\n\nConventions that apply to every endpoint:\n\n* **Identifiers are uuids.** Integer primary keys never appear in a URL, a\n  response body, an export or a log. The one exception is `/c/{token}`, which\n  carries a capability rather than a reference.\n* **Lists are cursor-paginated.** `?limit=&cursor=&sort=`. Offset pagination\n  skips or repeats rows when the underlying set changes between pages.\n* **Unknown filters are rejected**, never ignored.\n* **403 means visible but not permitted.** Anything outside your scope is 404 -\n  a 403 would confirm it exists.\n",
    "contact": {
      "name": "Privacy Office",
      "email": "privacy@example.org"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/notifications": {
      "get": {
        "tags": [
          "dashboard"
        ],
        "summary": "Notifications",
        "description": "Derived from the audit trail rather than a separate table.\n\nThere is no notifications table among the 22, and deriving the feed means it\ncan never disagree with the record it is describing.",
        "operationId": "notifications_notifications_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Notifications Notifications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/notifications/{log_uuid}/resend": {
      "post": {
        "tags": [
          "dashboard"
        ],
        "summary": "Resend",
        "description": "Re-deliver a failed notification.\n\nRestricted to DPO and DCO: re-sending a consent receipt puts a message in\nsomebody's inbox, and that is not an action a general user should be able to\ntrigger for an arbitrary event.",
        "operationId": "resend_notifications__log_uuid__resend_post",
        "parameters": [
          {
            "name": "log_uuid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Log Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Acknowledged"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Acknowledged": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Acknowledged",
        "description": "For state changes whose only interesting output is that they happened."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}